+++ CHANGELOG.md
@@ -8,11 +8,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ### Added
 
+- Add `msgpack-value` crate
+
 ### Fixed
 
 ### Removed
 
 ### Changed
+
+- Add more tests
 
 ---